Removed unused codes and refactoring
