+++ tools/spec-gen-sdk/package.json
@@ -37,9 +37,9 @@
     "@octokit/auth-app": "^2.4.5",
     "@octokit/core": "^4.0.5",
     "@octokit/rest": "18.0.3",
-    "@ts-common/azure-js-dev-tools": "^22.4.24",
     "ajv": "^6.12.6",
     "autorest": "3.6.1",
+    "commonmark": "0.31.2",
     "convict": "^6.2.4",
     "difflib": "^0.2.4",
     "dotenv": "^8.2.0",
