+++ tools/spec-gen-sdk/src/index.ts
@@ -1,8 +1,1 @@
-export { dotnet as csharp } from './langSpecs/langs/dotnet';
-export { go } from './langSpecs/langs/go';
-export { java } from './langSpecs/langs/java';
-export { javascript } from './langSpecs/langs/javascript';
-export * from './langSpecs/languageConfiguration';
-export { python } from './langSpecs/langs/python';
 export * from './sdkAutomation';
-export * from './specificationPREvent';
