--- tools/spec-gen-sdk/src/langSpecs/langs/dotnetTrack2.ts
@@ -1,17 +1,0 @@
-import { LanguageConfiguration } from '../languageConfiguration';
-import { getPackageName, createNugetPackage, findPackage, getInstallationInstructions } from './dotnet';
-
-/**
- * A language configuration for C#.
- */
-export const dotnetTrack2: LanguageConfiguration = {
-  name: '.NET Track2',
-  generatorPackageName: '@autorest/csharp-v3',
-  aliases: ['NET-Track2'],
-  packageRootFileName: /.*\.sln|^src$/,
-  packageNameCreator: getPackageName,
-  afterGenerationCommands: createNugetPackage,
-  packageCommands: findPackage,
-  packageNameAltPrefix: 'track2_',
-  installationInstructions: getInstallationInstructions
-};
