--- tools/spec-gen-sdk/src/tools/index.ts
@@ -1,22 +1,0 @@
-// Note: Export all tools
-
-export {
-    copyFolder,
-    createFolder,
-    deleteFolder,
-    copyFile,
-    getChildFilePaths
-} from './fileSystem2';
-
-export {
-    joinPath,
-    pathRelativeTo,
-    getPathName
-} from './path';
-
-export { replaceAll } from "./common";
-
-export {where} from "./arrays";
-
-export {mvnExecutable} from './mvn'
-
