+++ tools/spec-gen-sdk/src/utils/arrays.ts
@@ -1,10 +1,4 @@
 /**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for
- * license information.
- */
-
-/**
  * Get whether or not the provided array contains any values.
  * @param values The array to check.
  * @returns Whether or not the provided array contains any values.
@@ -21,12 +15,24 @@ export function any<T>(values: T[] | undefined): values is T[] {
  * @returns The array of values from the original values that match the provided condition.
  */
 export function where<T>(values: T[], condition: (value: T) => boolean): T[] {
-    const result: T[] = [];
-    for (const value of values) {
-      if (condition(value)) {
-        result.push(value);
-      }
+  const result: T[] = [];
+  for (const value of values) {
+    if (condition(value)) {
+      result.push(value);
     }
-    return result;
   }
+  return result;
+}
+
+/**
+ * Ensure that a value that is either a single value or an array is an array by wrapping single
+ * values in an array.
+ * @param value The value to ensure is an array.
+ * @param conversion The function that will be used to convert the non-array value to an array. This
+ * defaults to just creating a new array with the single value.
+ * @returns The array value.
+ */
+export function toArray<T>(value: T | T[], conversion: (valueToConvert: T) => T[] = (valueToConvert: T) => [valueToConvert]): T[] {
+  return value instanceof Array ? value : conversion(value);
+}
